- moved custom error classes to errors folder - applied new classes custom errors
