+++ app/errors/user_already_exist.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module UserAlreadyExist
+  class Error < StandardError
+    attr_reader :code
+
+    def initialize(message = "User account already exists.")
+      super(message)
+      @code = "user_already_exists"
+    end
+  end
+end
