+++ app/services/okta_auth_service.rb
@@ -40,8 +40,8 @@ class OktaAuthService < BaseAuthService
       create_user = user_atts["create_user"]
       users = User.where(primary_key => user_atts[primary_key])
 
-      raise self::UserAlreadyExists if create_user && !users.empty?
-      raise self::UserNotFound if !create_user.nil? && create_user && users.empty?
+      raise ::UserAlreadyExist::Error if create_user && !users.empty?
+      raise ::UserNotFound::Error if !create_user.nil? && create_user && users.empty?
 
       user = new_user(user_atts, primary_key) unless create_user.nil?
 
@@ -86,24 +86,6 @@ class OktaAuthService < BaseAuthService
             detail: detail
           }
         ]}
-    end
-  end
-
-  class UserNotFound < StandardError
-    attr_reader :code
-
-    def initialize(message = "User account not found.")
-      super(message)
-      @code = "user_not_found"
-    end
-  end
-
-  class UserAlreadyExists < StandardError
-    attr_reader :code
-
-    def initialize(message = "User account already exists.")
-      super(message)
-      @code = "user_already_exists"
     end
   end
 end
